requirements/requirements-test.lock.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/requirements-test.lock.txt requirements/requirements-base.lock.txt requirements/requirements-test.in
#
alabaster==0.7.16
    # via sphinx
anyio==4.4.0
    # via
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via
    #   -r requirements/requirements-base.lock.txt
    #   django
    #   django-stubs
astroid==3.2.4
    # via pylint
babel==2.15.0
    # via sphinx
backports-tarfile==1.2.0
    # via jaraco-context
build==1.2.1
    # via -r requirements/requirements-test.in
cachetools==5.4.0
    # via tox
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via cmarkgfm
chardet==5.2.0
    # via tox
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
cmarkgfm==2024.1.14
    # via readme-renderer
colorama==0.4.6
    # via
    #   sphinx-autobuild
    #   tox
coreapi==2.3.3
    # via -r requirements/requirements-test.in
coreschema==0.0.4
    # via coreapi
coverage[toml]==7.6.0
    # via pytest-cov
dill==0.3.8
    # via pylint
distlib==0.3.8
    # via virtualenv
django==5.0.7
    # via
    #   -r requirements/requirements-base.lock.txt
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
django-stubs==5.0.2
    # via
    #   -r requirements/requirements-test.in
    #   djangorestframework-stubs
django-stubs-ext==5.0.2
    # via django-stubs
djangorestframework==3.15.2
    # via -r requirements/requirements-base.lock.txt
djangorestframework-stubs==3.15.0
    # via -r requirements/requirements-test.in
doc8==1.1.1
    # via -r requirements/requirements-test.in
docutils==0.20.1
    # via
    #   -r requirements/requirements-test.in
    #   doc8
    #   readme-renderer
    #   restructuredtext-lint
    #   sphinx
    #   sphinx-jinja
    #   sphinx-rtd-theme
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
filelock==3.15.4
    # via
    #   tox
    #   virtualenv
flake8==7.1.0
    # via
    #   -r requirements/requirements-test.in
    #   flake8-comprehensions
    #   flake8-debugger
    #   flake8-isort
    #   flake8-print
    #   flake8-pyproject
    #   flake8-tuple
flake8-comprehensions==3.15.0
    # via -r requirements/requirements-test.in
flake8-debugger==4.1.2
    # via -r requirements/requirements-test.in
flake8-isort==6.1.1
    # via -r requirements/requirements-test.in
flake8-plugin-utils==1.3.3
    # via flake8-pytest-style
flake8-print==5.0.0
    # via -r requirements/requirements-test.in
flake8-pyproject==1.2.3
    # via -r requirements/requirements-test.in
flake8-pytest-style==2.0.0
    # via -r requirements/requirements-test.in
flake8-tuple==0.4.1
    # via -r requirements/requirements-test.in
h11==0.14.0
    # via uvicorn
idna==3.7
    # via
    #   anyio
    #   requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.0.0
    # via
    #   keyring
    #   twine
iniconfig==2.0.0
    # via pytest
isort[pyproject]==5.13.2
    # via
    #   -r requirements/requirements-test.in
    #   flake8-isort
    #   pylint
itypes==1.2.0
    # via coreapi
jaraco-classes==3.4.0
    # via keyring
jaraco-context==5.3.0
    # via keyring
jaraco-functools==4.0.1
    # via keyring
jinja2==3.1.4
    # via
    #   coreschema
    #   sphinx
    #   sphinx-jinja
keyring==25.2.1
    # via twine
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.3.0
    # via
    #   jaraco-classes
    #   jaraco-functools
mypy==1.11.0
    # via -r requirements/requirements-test.in
mypy-extensions==1.0.0
    # via mypy
nh3==0.2.18
    # via readme-renderer
packaging==24.1
    # via
    #   build
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
pbr==6.0.0
    # via stevedore
pkginfo==1.10.0
    # via twine
platformdirs==4.2.2
    # via
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   pytest
    #   tox
pycodestyle==2.12.0
    # via
    #   flake8
    #   flake8-debugger
    #   flake8-print
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via
    #   -r requirements/requirements-test.in
    #   doc8
    #   readme-renderer
    #   rich
    #   sphinx
pyjwt==2.8.0
    # via -r requirements/requirements-test.in
pylint==3.2.6
    # via -r requirements/requirements-test.in
pyproject-api==1.7.1
    # via tox
pyproject-hooks==1.1.0
    # via build
pytest==8.3.1
    # via
    #   -r requirements/requirements-test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/requirements-test.in
pytest-django==4.8.0
    # via -r requirements/requirements-test.in
readme-renderer[md]==43.0
    # via
    #   -r requirements/requirements-test.in
    #   twine
requests==2.32.3
    # via
    #   coreapi
    #   djangorestframework-stubs
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==1.0.0
    # via twine
restructuredtext-lint==1.4.0
    # via doc8
rfc3986==2.0.0
    # via twine
rich==13.7.1
    # via twine
six==1.16.0
    # via flake8-tuple
sniffio==1.3.1
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.4.7
    # via
    #   -r requirements/requirements-test.in
    #   sphinx-autobuild
    #   sphinx-jinja
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-autobuild==2024.4.16
    # via -r requirements/requirements-test.in
sphinx-jinja==2.0.2
    # via -r requirements/requirements-test.in
sphinx-rtd-theme==2.0.0
    # via -r requirements/requirements-test.in
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.6
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.8
    # via sphinx
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
sqlparse==0.5.1
    # via
    #   -r requirements/requirements-base.lock.txt
    #   django
starlette==0.38.0
    # via sphinx-autobuild
stevedore==5.2.0
    # via doc8
tomli==2.0.1
    # via
    #   build
    #   coverage
    #   django-stubs
    #   doc8
    #   flake8-pyproject
    #   mypy
    #   pylint
    #   pyproject-api
    #   pytest
    #   sphinx
    #   tox
tomlkit==0.13.0
    # via pylint
tox==4.16.0
    # via -r requirements/requirements-test.in
twine==5.1.1
    # via -r requirements/requirements-test.in
types-pyyaml==6.0.12.20240311
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.32.0.20240712
    # via djangorestframework-stubs
typing-extensions==4.12.2
    # via
    #   -r requirements/requirements-base.lock.txt
    #   anyio
    #   asgiref
    #   astroid
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
    #   uvicorn
uritemplate==4.1.1
    # via coreapi
urllib3==2.2.2
    # via
    #   requests
    #   twine
    #   types-requests
uvicorn==0.30.3
    # via sphinx-autobuild
virtualenv==20.26.3
    # via tox
watchfiles==0.22.0
    # via sphinx-autobuild
websockets==12.0
    # via sphinx-autobuild
wheel==0.43.0
    # via -r requirements/requirements-test.in
zipp==3.19.2
    # via importlib-metadata
